+++ src/inventory.h
@@ -20,12 +20,13 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #ifndef INVENTORY_HEADER
 #define INVENTORY_HEADER
 
-#include <sstream>
-#include <string>
-#include <vector>
-#include "irrlichttypes.h"
 #include "debug.h"
 #include "itemdef.h"
+#include "irrlichttypes.h"
+#include <istream>
+#include <ostream>
+#include <string>
+#include <vector>
 
 struct ToolCapabilities;
 
@@ -39,16 +40,12 @@ struct ItemStack
 
 	// Serialization
 	void serialize(std::ostream &os) const;
+	// Deserialization.  Pass itemdef unless you don't want aliases resolved.
 	void deSerialize(std::istream &is, IItemDefManager *itemdef = NULL);
 	void deSerialize(const std::string &s, IItemDefManager *itemdef = NULL);
 
 	// Returns the string used for inventory
-	std::string getItemString() const
-	{
-		std::ostringstream os(std::ios::binary);
-		serialize(os);
-		return os.str();
-	}
+	std::string getItemString() const;
 
 	/*
 		Quantity methods
